Rename exchange-connector to download
