+++ src/download/exchangeDirectoryParser.ts
