+++ test/download/resources/getAsset.json
