+++ test/download/resources/getAssetWithVersion.json
