+++ test/download/resources/with-category/raml-folder/exchange.json
