+++ test/download/resources/with-no-cc-category/raml-folder/exchange.json
